+++ src/components/utilities/Setup/DefenseSetup.vue
@@ -122,7 +122,42 @@ const selectedDefense = ref<UserDataStoreDefenseInterface|null>(null)
 
 const setupDefenses = computed(() => defenses.value.filter((defense) => defenseSetup.value.defenses[defense.incrementId] !== undefined))
 const defenseSelection = computed(() => defenses.value.filter((defense) => defense.userData && defenseSetup.value.defenses[defense.incrementId] === undefined))
-const totalDu = computed((): number => setupDefenses.value.reduce((accumulator, defense: UserDataStoreDefenseInterface) => accumulator + ((defense.defenseData?.defenseUnits ?? 0) * props.defenseSetup.defenses[defense.incrementId].defenseCount), 0))
+const totalDu = computed((): number => setupDefenses.value.reduce((accumulator: number, defense: UserDataStoreDefenseInterface) => {
+  if (!defense.defenseData) {
+    return accumulator
+  }
+
+  let defenseUnits: number = defense.defenseData.defenseUnits;
+  let defenseCount: number = (props.defenseSetup as UserDefenseSetupInterface).defenses[defense.incrementId].defenseCount;
+
+  if (defenseCount < 1) {
+    return accumulator
+  }
+
+  if (defense.defenseData.id === 'ProtonBeam') {
+    accumulator += defenseUnits;
+    defenseCount -= 1;
+    defenseUnits = 10;
+  }
+
+  if (defense.defenseData.id === 'WeaponManufacturer') {
+    if (defenseCount === 1) {
+      return accumulator + defenseUnits
+    }
+
+    // Weapon Manufacturer's main node
+    accumulator += defenseUnits;
+
+    // Weapon Manufacturer's second node
+    accumulator += 10;
+
+    // Weapon Manufacturer's remaining nodes
+    defenseUnits = 20;
+    defenseCount -= 2;
+  }
+
+  return accumulator + defenseUnits * defenseCount
+}, 0))
 const totalDps = computed((): number => {
   let calculatedTotalDps = 0